#  Mark json data as readonly and remove the need to slice it while reading