Also removed unnecessary Depth check and cached InObject (turning it into a field): 
https://github.com/dotnet/corefxlab/pull/2483#discussion_r218522745
https://github.com/dotnet/corefxlab/pull/2483#discussion_r218519322

~5% improvement when data is "compact"

+ Changed stack from int to bool to save space. No need to store 4 bytes (int), when 1 is sufficient (bool).

``` ini

BenchmarkDotNet=v0.11.1, OS=Windows 10.0.17744
Intel Core i7-6700 CPU 3.40GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100-alpha1-009460
  [Host]     : .NET Core 2.1.3 (CoreCLR 4.6.26725.06, CoreFX 4.6.26725.05), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.3 (CoreCLR 4.6.26725.06, CoreFX 4.6.26725.05), 64bit RyuJIT


```

**Before:**

|                               Method | IsDataCompact |        TestCase |          Mean |          Error |         StdDev | Allocated |
|------------------------------------- |-------------- |---------------- |--------------:|---------------:|---------------:|----------:|
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |       **BasicJson** |     **387.54 ns** |      **6.2077 ns** |      **5.8067 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |   **BasicLargeNum** |     **596.10 ns** |      **8.7268 ns** |      **8.1631 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |       **BroadTree** |   **7,469.65 ns** |     **98.1945 ns** |     **87.0468 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |        **DeepTree** |   **3,112.37 ns** |     **16.3954 ns** |     **12.8004 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |     **FullSchema1** |   **1,372.18 ns** |     **26.6702 ns** |     **24.9473 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |     **FullSchema2** |   **1,830.70 ns** |     **34.7344 ns** |     **37.1654 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |      **HelloWorld** |      **57.73 ns** |      **0.9430 ns** |      **0.8359 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |        **Json400B** |     **597.69 ns** |      **7.6057 ns** |      **7.1143 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |       **Json400KB** | **632,769.16 ns** | **12,332.3930 ns** | **14,680.8366 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |        **Json40KB** |  **58,137.78 ns** |  **1,153.4816 ns** |  **2,194.6187 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |         **Json4KB** |   **4,346.46 ns** |     **36.6541 ns** |     **32.4929 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |   **LotsOfNumbers** |   **2,790.77 ns** |     **37.2210 ns** |     **34.8166 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |   **LotsOfStrings** |     **937.60 ns** |     **17.5883 ns** |     **18.0619 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** | **ProjectLockJson** | **228,159.13 ns** |  **2,278.1924 ns** |  **1,902.3937 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |  **SpecialNumForm** |   **1,330.42 ns** |     **25.8794 ns** |     **25.4171 ns** |       **0 B** |

**After:**

|                               Method | IsDataCompact |        TestCase |          Mean |          Error |         StdDev | Allocated |
|------------------------------------- |-------------- |---------------- |--------------:|---------------:|---------------:|----------:|
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |       **BasicJson** |     **367.59 ns** |      **7.2716 ns** |      **8.0824 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |   **BasicLargeNum** |     **556.06 ns** |      **7.4098 ns** |      **6.9311 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |       **BroadTree** |   **7,015.26 ns** |     **40.8206 ns** |     **34.0871 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |        **DeepTree** |   **3,096.43 ns** |     **59.3949 ns** |     **68.3992 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |     **FullSchema1** |   **1,272.49 ns** |     **12.0228 ns** |     **11.2462 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |     **FullSchema2** |   **1,794.53 ns** |     **34.0180 ns** |     **33.4102 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |      **HelloWorld** |      **58.19 ns** |      **0.7608 ns** |      **0.6744 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |        **Json400B** |     **581.46 ns** |     **11.3057 ns** |     **13.4587 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |       **Json400KB** | **610,963.23 ns** | **22,499.5760 ns** | **22,097.5899 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |        **Json40KB** |  **55,488.30 ns** |  **1,825.6843 ns** |  **3,245.1520 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |         **Json4KB** |   **4,110.18 ns** |     **77.1468 ns** |     **72.1632 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |   **LotsOfNumbers** |   **2,716.84 ns** |     **41.8256 ns** |     **39.1237 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |   **LotsOfStrings** |     **911.44 ns** |      **3.6986 ns** |      **2.8876 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** | **ProjectLockJson** | **213,966.40 ns** |  **1,168.3763 ns** |    **912.1913 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |  **SpecialNumForm** |   **1,246.48 ns** |      **8.2556 ns** |      **6.8938 ns** |       **0 B** |